Move user list component to shared, add app header
